Moved markdown methods to separate object
